Remove OneHotEncoder and rename OneHotEncoderEstimator to OneHotEncoder.
